refact(All apis):split all the test cases into separate file
