+++ src/modules/Awards/deleteAwards/deleteAwards.test.js
@@ -1,0 +1,103 @@
+import mongoose from 'mongoose';
+import 'dotenv/config'; 
+import request from 'supertest';
+import express from 'express';
+import awardsRoutes from '../routes';
+
+const MongoDbString = process.env.MONGODBSTRING;
+
+mongoose.connect(MongoDbString);
+
+const db = mongoose.connection;
+
+db.on('error', console.error.bind(console, 'MongoDB connection error:'));
+db.once('open', () => {
+  console.log('Connected to MongoDB');
+});
+
+const app = express();
+app.use(express.json());
+app.use('/', awardsRoutes);
+
+
+
+// Delete awards api
+
+describe('Awards Routes - Delete User Awards', () => {
+  it('should return success response for valid awards deletion', async () => {
+    const validAwardsToDelete = {
+      awards: ["65f003041ea36b47123ad212"] 
+    };
+
+    const response = await request(app)
+      .delete('/deleteUserAwards')
+      .set('Authorization', 'Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJwaG9uZU51bWJlciI6Iis5MTk3ODc1NDYzMzUiLCJpYXQiOjE3MTA1MDA4ODgsImV4cCI6MTcxMTM2NDg4OH0.76LxDkKSpAup4rsdm0uxblh1NP5zy7tyiiyzG79BFdk')
+      .send(validAwardsToDelete);
+
+    expect(response.statusCode).toBe(200);
+    expect(response.body).toHaveProperty('success', true);
+    expect(response.body).toHaveProperty('isAuth', true);
+    expect(response.body).toHaveProperty('message', 'Awards Deleted Successfully.');
+  });
+
+  it('should return error response for missing authorization token', async () => {
+    const validAwardsToDelete = {
+      awards: ["65f003041ea36b47123ad212"] 
+    };
+
+    const response = await request(app)
+      .delete('/deleteUserAwards')
+      .send(validAwardsToDelete);
+
+    expect(response.statusCode).toBe(500);
+    expect(response.body).toHaveProperty('success', false);
+    expect(response.body).toHaveProperty('message', 'User not authorized');
+  });
+
+  it('should return error response for invalid awards', async () => {
+    const invalidAwardsToDelete = {
+      awards: [""] 
+    };
+
+    const response = await request(app)
+      .delete('/deleteUserAwards')
+      .set('Authorization', 'Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJwaG9uZU51bWJlciI6Iis5MTk3ODc1NDYzMzUiLCJpYXQiOjE3MTA1MDA4ODgsImV4cCI6MTcxMTM2NDg4OH0.76LxDkKSpAup4rsdm0uxblh1NP5zy7tyiiyzG79BFdk') 
+      .send(invalidAwardsToDelete);
+
+    expect(response.statusCode).toBe(500);
+    expect(response.body).toHaveProperty('success', false);
+    expect(response.body).toHaveProperty('errors');
+    expect(response.body.errors.length).toBeGreaterThan(0);
+  });
+  it('should return error response for invalid awards', async () => {
+    const invalidAwardsToDelete = {
+      awards: [""] ,
+      age:"dvsdfd"
+    };
+
+    const response = await request(app)
+      .delete('/deleteUserAwards')
+      .set('Authorization', 'Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJwaG9uZU51bWJlciI6Iis5MTk3ODc1NDYzMzUiLCJpYXQiOjE3MTA1MDA4ODgsImV4cCI6MTcxMTM2NDg4OH0.76LxDkKSpAup4rsdm0uxblh1NP5zy7tyiiyzG79BFdk') 
+      .send(invalidAwardsToDelete);
+
+    expect(response.statusCode).toBe(500);
+    expect(response.body).toHaveProperty('success', false);
+    expect(response.body).toHaveProperty('errors');
+    expect(response.body.errors.length).toBeGreaterThan(0);
+  });
+  it('should return error response for invalid awards', async () => {
+    const invalidAwardsToDelete = {
+      awards: ""
+    };
+
+    const response = await request(app)
+      .delete('/deleteUserAwards')
+      .set('Authorization', 'Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJwaG9uZU51bWJlciI6Iis5MTk3ODc1NDYzMzUiLCJpYXQiOjE3MTA1MDA4ODgsImV4cCI6MTcxMTM2NDg4OH0.76LxDkKSpAup4rsdm0uxblh1NP5zy7tyiiyzG79BFdk') 
+      .send(invalidAwardsToDelete);
+
+    expect(response.statusCode).toBe(500);
+    expect(response.body).toHaveProperty('success', false);
+    expect(response.body).toHaveProperty('errors');
+    expect(response.body.errors.length).toBeGreaterThan(0);
+  });
+});
